# Add preliminary performance tester