mado-perf is a standalone benchmarking program, modeled after x11perf, that evaluates the performance of compositing functions.

Expected output format:
       2000 reps @   0.0140 msec (  71523084.1/sec): 1x1 argb32 source
       2000 reps @   0.0140 msec (  71523084.1/sec): 1x1 argb32 source
       2000 reps @   0.0130 msec (  77032700.4/sec): 1x1 argb32 source
       ...
       2000 reps @   0.0140 msec (  71523084.1/sec): 1x1 argb32 source
      80000 trep @   0.5578 msec (  73726930.6/sec): 1x1 argb32 source 
 <div id='description'>
    <a href="https://bito.ai#summarystart"></a>
<h3>Summary by Bito</h3>
This pull request introduces 'mado-perf', a performance testing tool for benchmarking compositing functions, modeled after 'x11perf'. It includes updates to the Makefile and Kconfig for integration, along with various test cases for different pixel formats and sizes to ensure accurate performance metrics.
<!-- Disabling unit_tests and post_effort_to_review
<br>
<br>
<b>Unit tests added</b>: True
<br>
<br>
<b>Estimated effort to review (1-5, lower is better)</b>: 2
-->
</div>